backend/requirements.txt
fastapi==0.95.1
passlib==1.7.4
pydantic==1.10.7
PyJWT==2.6.0
PyJWT==2.6.0
SQLAlchemy==2.0.9
